refactor: Custom Api Response 디렉토리 이동
